requirements.txt
# LangChain & AI (updated versions)
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.0.8

# Vector Database
chromadb==0.4.24

# Document Processing
pypdf==4.2.0
python-docx==1.1.0

# API
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Frontend
streamlit==1.34.0

# Utils
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

# OpenAI (compatible version)
openai==1.30.1
